refactor: separate commands loader into a file
