+++ src/core/deployGuildCommands.ts
@@ -2,8 +2,7 @@ import dotenv from 'dotenv';
 import { Routes } from 'discord.js';
 import { REST } from '@discordjs/rest';
 import config from '#root/config';
-import readDirectory from '#libs/readDirectory';
-import { URL } from 'node:url';
+import getCommands from '#libs/getCommands';
 
 dotenv.config();
 
@@ -12,16 +11,7 @@ dotenv.config();
 
 	if (!botToken) return;
 
-	const commands = [];
-	const { directoryPath, filteredFiles } = await readDirectory('./../commands');
-
-	for (const file of filteredFiles) {
-		const filePath = new URL(`./commands/${file}`, directoryPath).href;
-		const { AenexCommand } = await import(filePath);
-		const command = new AenexCommand();
-
-		commands.push(command.buildApplicationCommand().toJSON());
-	}
+	const commands = await getCommands();
 
 	const rest = new REST({
 		version: '10',
